Fix Linux and HOL Install Scripts

+++ INTRODUCTION/install/INTRODUCTION.sh
@@ -518,5 +518,8 @@ SERVER=${5}
 demohome="$(dirname "$(pwd)")"
 logfilename=$demohome/install/INTRODUCTION.log
 echo "Log File : $logfilename"
-rm $logfilename
+if [ -f "$logfilename" ]
+then
+  rm $logfilename
+fi
 doInstall 2>&1 | tee -a $logfilename
